split components, better styiling
