+++ hat/assets/js/apps/Iaso/domains/instances/components/ActionTableColumnComponent.tsx
@@ -54,7 +54,7 @@ const getUrlInstance = data => {
     return `${initialUrl}`;
 };
 
-const ActionTableColumnComponent = ({ settings }) => {
+export const ActionTableColumnComponent = ({ settings }) => {
     const user = useCurrentUser();
     const [, , setFieldErrors] = useFormState(
         initialFormState(
