+++ hat/assets/js/apps/Iaso/domains/instances/components/SubmissionRail/SubmissionRail.tsx
@@ -8,10 +8,10 @@ import PlaceOutlinedIcon from '@mui/icons-material/PlaceOutlined';
 import ShieldOutlinedIcon from '@mui/icons-material/ShieldOutlined';
 import { Paper } from '@mui/material';
 import { useSafeIntl } from 'bluesquare-components';
+import { getInstancesFilesList } from 'Iaso/domains/instances/utils';
+import { numericValues } from 'Iaso/domains/instances/utils/intl';
 import MESSAGES from '../../messages';
 import { Instance } from '../../types/instance';
-import { getInstancesFilesList } from '../../utils';
-import { numericValues } from '../../utils/intl';
 import InstanceDetailsChangeRequests from '../InstanceDetailsChangeRequests';
 import { InstanceDetailsExportRequestsContent } from '../InstanceDetailsExportRequests';
 import InstanceDetailsLocation from '../InstanceDetailsLocation';
